requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
psutil
GPUtil
wmi
pywin32
SQLAlchemy
werkzeug
websockets
setuptools
psycopg2-binary